chapter7/requirements.txt
langchain-core==0.1.18
langchain-openai==0.0.5
langchain-community==0.0.17
python-dotenv==1.0.1
urllib3==1.26.18
slack-bolt==1.18.1
boto3==1.34.34
momento==1.17.0
